docker/requirements.txt
selenium
webdriver-manager
python-dotenv
dateparser
sentence-transformers
transformers
torch
scikit-learn
numpy
pymongo
pandas
plotly
altair
matplotlib
pytz
dateparser
TorchCRF
jsonlines
prettytable
tensorboardX
vncorenlp
apache-airflow-providers-mongo